Add files via upload
It's my setup and portfolio. I'm Ismail Irfan. I'm a Hafiz-e-Quran as well as intermediate certified through national college and certified through Aptech Institute in Information System Management. I'm a web developer.

It's by the grace of God that my school life was dedicated to prayers, aiming for the happiness and satisfaction of all human beings. I pray for the equity and satisfaction of all human beings.